% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RM.R
\name{ERM}
\alias{ERM}
\title{Generate Euclidean random matrix}
\usage{
ERM(N, dim = 3, f = function(x) {
    exp(-x)
})
}
\arguments{
\item{N}{the number of point}

\item{dim}{the dimension of the point}

\item{f}{the function from distance to correlation}
}
\value{
a Euclidean random matrix corr
}
\description{
Generate Euclidean random matrix
}
\examples{
\dontrun{
RM(100, 3, f = exp)
}
}
