% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_spectrum.R
\name{log_spectrum}
\alias{log_spectrum}
\title{computation log eigen spectrum of random matrix}
\usage{
log_spectrum(N, dim, n = 100, f = function(x) {
    exp(-x)
}, R = ERM)
}
\arguments{
\item{N}{the number of point}

\item{dim}{the dimension of the point}

\item{n}{number of repetitions}

\item{f}{the function from distance to correlation}

\item{R}{the function Generate random matrix}
}
\value{
eigen spectrum
}
\description{
computation log eigen spectrum of random matrix
}
\examples{
\dontrun{
x = eigen_spectrum(10, 3)$x
y = eigen_spectrum(10, 3)$y
plot(x,y)
}
}
