\name{StatComp21051-package}
\alias{StatComp21051-package}
\docType{package}
\title{
This package provides several functions for benchmarking
}
\description{
This project studies the scale invariance of random matrices, it includes a series of random matrix generation, eigenvalue spectrum, and plotting function.
}
\author{
Zezhen Wang
Maintainer: Zezhen Wang <wzz1999@mail.ustc.edu.cn>
}
\keyword{ package }